add unblockJob method and JobsService Client
